pyproject.toml
[tool.poetry]
name = "tracer_api"
version = "0.1.0"
description = ""
authors = ["wassafshahzad <wassafshahzad@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.2"
dj-database-url = "^2.2.0"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
pycountry = "^24.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
